refactor: split CLI into command modules and introduce clap
Split main.rs (5168 lines) into 9 command modules under
src/commands/. Replace hand-rolled argument parsing with clap
derive macros. Unify exit codes via Result<ExitCode> flow.
