+++ crates/mmd-anim-cli/src/commands/mod.rs
@@ -1,0 +1,8 @@
+pub(crate) mod bench;
+pub(crate) mod compare;
+pub(crate) mod export;
+pub(crate) mod golden;
+pub(crate) mod import;
+pub(crate) mod oracle;
+pub(crate) mod parse;
+pub(crate) mod patch;
